# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FleetSaaS school bus tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /passengers:
    get:
      operationId: listPassengers
      tags: [passengers]
      summary: List all passengers for today's trip checklist
      responses:
        "200":
          description: Passenger list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Passenger"
    post:
      operationId: createPassenger
      tags: [passengers]
      summary: Register a new passenger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PassengerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
  /passengers/{id}:
    delete:
      operationId: deletePassenger
      tags: [passengers]
      summary: Remove a passenger
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
    get:
      operationId: getPassenger
      tags: [passengers]
      summary: Get passenger by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Passenger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
    patch:
      operationId: updatePassenger
      tags: [passengers]
      summary: Update passenger profile or photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PassengerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
  /passengers/{id}/board:
    post:
      operationId: boardPassenger
      tags: [passengers]
      summary: Mark passenger as boarded (driver action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated passenger with boarding status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
  /passengers/{id}/unboard:
    post:
      operationId: unboardPassenger
      tags: [passengers]
      summary: Mark passenger as unboarded (driver action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated passenger with pending status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"
  /passengers/{id}/leave:
    post:
      operationId: markPassengerLeave
      tags: [passengers]
      summary: Mark passenger as on leave (parent or admin action)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated passenger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Passenger"

  # ─── TRIPS ──────────────────────────────────────────────────────────────────
...
  /routes:
    get:
      operationId: listRoutes
      tags: [routes]
      summary: List all routes for this tenant
      responses:
        "200":
          description: Route list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Route"
    post:
      operationId: createRoute
      tags: [routes]
      summary: Create a new route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteInput"
      responses:
        "201":
          description: Created route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
  /routes/{id}:
    patch:
      operationId: updateRoute
      tags: [routes]
      summary: Update route name or assigned driver/vehicle (fleet swap)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteUpdate"
      responses:
        "200":
          description: Updated route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Route"
    delete:
      operationId: deleteRoute
      tags: [routes]
      summary: Delete a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /routes/{id}/stations:
    get:
      operationId: listRouteStations
      tags: [routes]
      summary: Get ordered stations for a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ordered station list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteStation"
    post:
      operationId: addRouteStation
      tags: [routes]
      summary: Add a station to a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteStationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteStation"
  /routes/{id}/stations/{stationId}:
    delete:
      operationId: removeRouteStation
      tags: [routes]
      summary: Remove a station from a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: stationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /routes/{id}/stations/reorder:
    post:
      operationId: reorderRouteStations
      tags: [routes]
      summary: Reorder stations within a route
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderStationsInput"
      responses:
        "200":
          description: Reordered stations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RouteStation"

  # ─── GEOCODE ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Route:
      type: object
      required: [id, tenantId, name, isActive]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        driverName:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
    RouteInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        driverId:
          type: integer
        vehicleId:
          type: integer
    RouteUpdate:
      type: object
      properties:
        name:
          type: string
        driverId:
          type: ["integer", "null"]
        vehicleId:
          type: ["integer", "null"]
        isActive:
          type: boolean
    RouteStation:
      type: object
      required: [id, routeId, stationId, position]
      properties:
        id:
          type: integer
        routeId:
          type: integer
        stationId:
          type: integer
        position:
          type: integer
        stationName:
          type: string
        lat:
          type: number
        lng:
          type: number
        radius:
          type: integer
    RouteStationInput:
      type: object
      required: [stationId]
      properties:
        stationId:
          type: integer
        position:
          type: integer
    ReorderStationsInput:
      type: object
      required: [orderedStationIds]
      properties:
        orderedStationIds:
          type: array
          items:
            type: integer
...
    Passenger:
      type: object
      required: [id, name, role, status, stationId]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, staff]
        status:
          type: string
          enum: [pending, boarded, leave]
        stationId:
          type: integer
        routeId:
          type: ["integer", "null"]
        stationName:
          type: ["string", "null"]
        boardedAt:
          type: ["string", "null"]
          format: date-time
        liveToday:
          type: integer
          description: 1 if the passenger confirmed they are riding today, 0 otherwise
        quickMessage:
          type: ["string", "null"]
          description: Last quick status message sent by the passenger
    PassengerInput:
      type: object
      required: [name, role, stationId]
      properties:
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        role:
          type: string
          enum: [student, staff]
        stationId:
          type: integer
        routeId:
          type: integer
    PassengerUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        photoUrl:
          type: string
        stationId:
          type: integer
        routeId:
          type: ["integer", "null"]
        liveToday:
          type: integer
        quickMessage:
          type: string
...